0.9.0-alpha.1: migrate to API V3 and expand management

+++ app/src/iosMain/kotlin/crucible/lens/data/util/CryptoUtils.kt
@@ -10,7 +10,7 @@ import platform.CoreCrypto.CC_SHA256_DIGEST_LENGTH
 actual object PlatformCrypto {
     @OptIn(ExperimentalForeignApi::class)
     actual fun sha256Hex(bytes: ByteArray): String {
-        val digest = UByteArray(CC_SHA256_DIGEST_LENGTH.toInt())
+        val digest = UByteArray(CC_SHA256_DIGEST_LENGTH)
         bytes.usePinned { src ->
             digest.usePinned { dst ->
                 CC_SHA256(src.addressOf(0).reinterpret<kotlinx.cinterop.UByteVar>(), bytes.size.toUInt(), dst.addressOf(0))
